pyproject.toml
[tool.poetry]
name = "fullstats_test_task"
version = "0.1.0"
description = ""
authors = ["Timur <phoenixwq01@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2.0"
dj-database-url = "^0.5.0"
djangorestframework = "^3.13.1"
djoser = "^2.1.0"
python-dotenv = "^0.20.0"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
